overload and template draw method for easier extending
